+++ apps/meteor/app/e2e/server/methods/setUserPublicAndPrivateKeys.ts
@@ -1,11 +1,11 @@
-import { Users } from '@rocket.chat/models';
+import { Rooms, Users } from '@rocket.chat/models';
 import type { ServerMethods } from '@rocket.chat/ui-contexts';
 import { Meteor } from 'meteor/meteor';
 
 declare module '@rocket.chat/ui-contexts' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
 	interface ServerMethods {
-		'e2e.setUserPublicAndPrivateKeys'({ public_key, private_key }: { public_key: string; private_key: string }): void;
+		'e2e.setUserPublicAndPrivateKeys'({ public_key, private_key }: { public_key: string; private_key: string; force?: boolean }): void;
 	}
 }
 
@@ -19,9 +19,22 @@ Meteor.methods<ServerMethods>({
 			});
 		}
 
+		if (!keyPair.force) {
+			const keys = await Users.fetchKeysByUserId(userId);
+
+			if (keys.private_key && keys.public_key) {
+				throw new Meteor.Error('error-keys-already-set', 'Keys already set', {
+					method: 'e2e.setUserPublicAndPrivateKeys',
+				});
+			}
+		}
+
 		await Users.setE2EPublicAndPrivateKeysByUserId(userId, {
 			private_key: keyPair.private_key,
 			public_key: keyPair.public_key,
 		});
+
+		const subscribedRoomIds = await Rooms.getSubscribedRoomIdsWithoutE2EKeys(userId);
+		await Rooms.addUserIdToE2EEQueueByRoomIds(subscribedRoomIds, userId);
 	},
 });