+++ apps/meteor/client/components/message/variants/thread/ThreadMessageContent.tsx
@@ -1,5 +1,6 @@
 import type { IThreadMainMessage, IThreadMessage } from '@rocket.chat/core-typings';
 import { isE2EEMessage, isQuoteAttachment } from '@rocket.chat/core-typings';
+import { MessageBody } from '@rocket.chat/fuselage';
 import type { TranslationKey } from '@rocket.chat/ui-contexts';
 import { useSetting, useUserId, useTranslation } from '@rocket.chat/ui-contexts';
 import type { ReactElement } from 'react';
@@ -37,12 +38,16 @@ const ThreadMessageContent = ({ message }: ThreadMessageContentProps): ReactElem
 
 	const normalizedMessage = useNormalizedMessage(message);
 
+	const isMessageEncrypted = encrypted && normalizedMessage?.e2e === 'pending';
+
 	const quotes = normalizedMessage?.attachments?.filter(isQuoteAttachment) || [];
 
 	const attachments = normalizedMessage?.attachments?.filter((attachment) => !isQuoteAttachment(attachment)) || [];
 
 	return (
 		<>
+			{isMessageEncrypted && <MessageBody>{t('E2E_message_encrypted_placeholder')}</MessageBody>}
+
 			{!!quotes?.length && <Attachments attachments={quotes} />}
 
 			{!normalizedMessage.blocks?.length && !!normalizedMessage.md?.length && (
@@ -50,7 +55,6 @@ const ThreadMessageContent = ({ message }: ThreadMessageContentProps): ReactElem
 					{(!encrypted || normalizedMessage.e2e === 'done') && (
 						<MessageContentBody md={normalizedMessage.md} mentions={normalizedMessage.mentions} channels={normalizedMessage.channels} />
 					)}
-					{encrypted && normalizedMessage.e2e === 'pending' && t('E2E_message_encrypted_placeholder')}
 				</>
 			)}