+++ packages/core-typings/src/IRoom.ts
@@ -94,7 +94,7 @@ export interface IRoom extends IRocketChatRecord {
 	/* @deprecated */
 	customFields?: Record<string, any>;
 
-	channel?: { _id: string };
+	usersWaitingForE2EKeys?: { userId: IUser['_id']; ts: Date }[];
 }
 
 export const isRoomWithJoinCode = (room: Partial<IRoom>): room is IRoomWithJoinCode =>